Subtracing needs factorized spins
The ordering of the states is indeed a problem: if we do not generate the states (and so the indices) in the correct order (i.e. one that is consistent with the original buildup of the dm) we trace over the wrong elements.

So here we factorize the up and dw impurity sectors in the <subtrace()> subroutine. With Nlat=2, tracing to the local one, we get the diagonal correctly up to machine precision (and so the semi-analytical benchmars **appears** to pass), but we still have off-diagonal terms, which are **dependent on the size of the bath** -> this points for sure to an error in the trace, since whatever reduced density matrix cannot depend on degrees of freedom that are already traced out. Right?