INTERMEDIATE: incremental subtracing
Testing incremental subtracing: starting with a high Nlat (*) we trace k=1,2,3 sites and look at the resulting RDMs. They have the right dimension [and the local one matches with the Norb=1 benchmark].

(*) Nlat=4 for Norb=1; Nlat=3 for Norb=2.

TODO: write a "recursive" version of the <subtrace()> subroutine, that takes a **generic** matrix (allocatable, not specified size) and traces away **just one site**
> That way we would be able to check all the reduced matrices to be correct.