STABLE: incremental subtracing succesfully tested
Added <sitetrace()>, which takes in a generic dimension density matrix and traces out a 1-site-reduced dm.

Tested with the current input-file: site-by-site tracing (with the new subroutine) and incremental tracing with <subtrace()> give the same dm, rank-by-rank, to machine precision.

>> SUBTRACING ALGORITHM can be trusted completely.

--------

TODO: maybe we should merge <sitetrace()> and <subtrace()> to have a totally generic subroutine taking in a "big" dm, with the corresponding number of sites and returning a "small" dm, reduced by an abritrary number of sites. This would be the definitive subroutine to be exported to https://github.com/QcmPlab/CDMFT-LANC-ED